User can now add and remove keywords using command line.
The counter doesn't seem to be blocking any more. Actually the blocking part was the print to the console and has been removed.

Next: Find a way for the leader board to have access to all keywords.
By writing it into the database ?? Or having some kind of global value for streaming?
Both sound quite logical.

FIXME : BTW, counter is also still working on a static list of keywords placed in data.