Solves bug :
It is possible to create the same hashtag several times in the databse. => Add existence checks

A bug should not be created twice any more in the trendy list.

FIXME:
Fix waas done in the plane with no internet connexion nor ressource.
Code is really bad and not pythonic :s.
Think about doing better.

database has to be cleaned from hashtags to remove duplicates now.